Allow using IFD to generate attribute names (e.g. `package.${system}.hello`)
This in addition also strips away unset attributes set using `flake`.

The core issue with IFD in this context, is that we want to strip away unused attributes.
AFAICT, flake-parts did this in a buggy way before. Removing this entirely while
fixing support for IFD was easier, but to keep support for filtering away unset
attributes, a bit more complexity was needed.

The issue is that we can't use `null` as a special value to mean "unset".
The reason is that while filtering over `allSystems.${system}`, checking
whether `packages` has been set, doing `null == packages` will cause `packages`,
an attribute set, to be evaluated to WHNF. WHNF for an attribute set entails
evaluating the _keys_, and if the keys are generated using IFD, then this will
fail, even if we know that no matter what the key is, this check will return false.

If the Nix evaluator is ever fixed, this feature can be somewhat cleaned up by using
null checks rather than checking `isDefined`.

To make use of this feature, you can not simply do
```nix
packages = import drv;
```

You must do
```nix
packages = mkOverride 100 (mkOrder 1000 (import drv));
```

The reason is that the module system internally will try to
check `._type`, but this of course will cause all of the other
keys to also be evaluated.

This would also be unnecessary if the Nix evaluator were fixed.

Closes #57